refactor(library): merge http back to platform (#16)
